Overriden the rest_auth UserDetails end point
Implemented a more context/app specific user details end point that returns data more relevant to Talespace. This will enable implementation of the user profile on the frontend.

Additionally, the user model is included in the same UserDetail endpoint as a nested object which should prevent the need for multiple calls to the server when rendering the users profile.